+++ src/nvim/api/vim.c
@@ -10,7 +10,7 @@
 #include "nvim/api/private/helpers.h"
 #include "nvim/api/private/defs.h"
 #include "nvim/api/buffer.h"
-#include "nvim/os/channel.h"
+#include "nvim/msgpack_rpc/channel.h"
 #include "nvim/os/provider.h"
 #include "nvim/vim.h"
 #include "nvim/buffer.h"
