+++ src/nvim/msgpack_rpc/channel.h
@@ -1,5 +1,5 @@
-#ifndef NVIM_OS_CHANNEL_H
-#define NVIM_OS_CHANNEL_H
+#ifndef NVIM_MSGPACK_RPC_CHANNEL_H
+#define NVIM_MSGPACK_RPC_CHANNEL_H
 
 #include <stdbool.h>
 #include <uv.h>
@@ -10,6 +10,6 @@
 #define METHOD_MAXLEN 512
 
 #ifdef INCLUDE_GENERATED_DECLARATIONS
-# include "os/channel.h.generated.h"
+# include "msgpack_rpc/channel.h.generated.h"
 #endif
-#endif  // NVIM_OS_CHANNEL_H
+#endif  // NVIM_MSGPACK_RPC_CHANNEL_H
