+++ src/nvim/msgpack_rpc/defs.h
@@ -1,19 +1,8 @@
-#ifndef NVIM_OS_MSGPACK_RPC_H
-#define NVIM_OS_MSGPACK_RPC_H
-
-#include <stdint.h>
+#ifndef NVIM_MSGPACK_RPC_DEFS_H
+#define NVIM_MSGPACK_RPC_DEFS_H
 
 #include <msgpack.h>
 
-#include "nvim/func_attr.h"
-#include "nvim/api/private/defs.h"
-#include "nvim/os/wstream.h"
-
-typedef enum {
-  kUnpackResultOk,        /// Successfully parsed a document
-  kUnpackResultFail,      /// Got unexpected input
-  kUnpackResultNeedMore   /// Need more data
-} UnpackResult;
 
 /// The rpc_method_handlers table, used in msgpack_rpc_dispatch(), stores
 /// functions of this type.
@@ -21,9 +10,8 @@ typedef Object (*rpc_method_handler_fn)(uint64_t channel_id,
                                         msgpack_object *req,
                                         Error *error);
 
-
 /// Initializes the msgpack-rpc method table
-void msgpack_rpc_init(void);
+void msgpack_rpc_init_method_table(void);
 
 void msgpack_rpc_init_function_metadata(Dictionary *metadata);
 
@@ -43,9 +31,4 @@ Object msgpack_rpc_dispatch(uint64_t channel_id,
                             Error *error)
   FUNC_ATTR_NONNULL_ARG(2) FUNC_ATTR_NONNULL_ARG(3);
 
-
-#ifdef INCLUDE_GENERATED_DECLARATIONS
-# include "os/msgpack_rpc.h.generated.h"
-#endif
-
-#endif  // NVIM_OS_MSGPACK_RPC_H
+#endif  // NVIM_MSGPACK_RPC_DEFS_H
