--- src/nvim/os/server_defs.h
@@ -1,7 +1,0 @@
-#ifndef NVIM_OS_SERVER_DEFS_H
-#define NVIM_OS_SERVER_DEFS_H
-
-typedef struct server Server;
-
-#endif  // NVIM_OS_SERVER_DEFS_H
-
