+++ src/nvim/os_unix.c
@@ -54,8 +54,8 @@
 #include "nvim/os/shell.h"
 #include "nvim/os/signal.h"
 #include "nvim/os/job.h"
-#include "nvim/os/msgpack_rpc.h"
-#include "nvim/os/msgpack_rpc_helpers.h"
+#include "nvim/msgpack_rpc/helpers.h"
+#include "nvim/msgpack_rpc/defs.h"
 
 #if defined(HAVE_SYS_IOCTL_H)
 # include <sys/ioctl.h>
@@ -166,8 +166,6 @@ void mch_init(void)
   mac_conv_init();
 #endif
 
-  msgpack_rpc_init();
-  msgpack_rpc_helpers_init();
   event_init();
 }
